Update Task to non-abstract and remove Todo
There is no need for a separate Todo class and a Task superclass as
we only have 1 type of task in our program with many attributes,
instead of many types of task.
