+++ src/test/java/seedu/duke/task/TaskTest.java
@@ -6,24 +6,24 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class TodoTest {
+class TaskTest {
 
     @Test
     void getDescription_inputDescription_returnsCorrectDescription() {
         String inputString = "test description";
-        Todo todo = new Todo(inputString);
+        Task todo = new Task(inputString);
         assertEquals(inputString, todo.getDescription());
     }
 
     @Test
     void getIsDone_isDone_returnsFalseByDefault() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         assertFalse(todo.getIsDone());
     }
 
     @Test
     void markAsDone_setIsDone_todoSetAsDone() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         assertFalse(todo.getIsDone());
         todo.markAsDone();
         assertTrue(todo.getIsDone());
@@ -31,7 +31,7 @@ class TodoTest {
 
     @Test
     void toFile_getToFile_returnsCorrectString() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         String fileString = todo.toFile();
         String expectedString = "T | 0 | test description | 0 |  | ";
         assertEquals(expectedString, fileString);
@@ -39,7 +39,7 @@ class TodoTest {
 
     @Test
     void testToString_toString_returnsCorrectString() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         String todoString = todo.toString();
         String expectedString = "[T][N] test description (p:0)";
         assertEquals(expectedString, todoString);
@@ -47,13 +47,13 @@ class TodoTest {
 
     @Test
     void todo_noInputPriority_setsDefaultPriority() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         assertEquals(0, todo.getPriority());
     }
 
     @Test
     void setPriority_validPriority_setsCorrectPriority() {
-        Todo todo = new Todo("test description");
+        Task todo = new Task("test description");
         todo.setPriority(5);
         assertEquals(5, todo.getPriority());
     }
